requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
python-multipart
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv
tenacity
python-dateutil
pyotp
qrcode
lxml